move accumulator and columnar value
